Restructured slightly by putting the function in pad_util instead
