+++ pad_api_data/pad_etl/common/dungeon_types.py
@@ -10,7 +10,7 @@ DUNGEON_TYPE = {
 
 
 # These for sure might need a rename/reclassification. I gave them as best a designation as I could
-DUNGEON_TYPE_DESCRIPTORS = {
+DUNGEON_TYPE_COMMENTS = {
     0: 'Normal',  # Something to do with them at least
     1: 'Special Descended',
     2: 'Endless',
@@ -56,22 +56,3 @@ DUNGEON_TYPE_DESCRIPTORS = {
     600011: 'Multiplayer',
 }
 
-# Simply returns a comment for a input raw value. Doing it this way is necessary as to not have to put all the values
-# directly into a dictionary when multiple val's correspond to a single comment, but are unnecessarily delineated
-def get_dungeon_comment(val):
-    if val in range(5611, 5615):
-        return "Retired Special Dungeons"  # These are the last normal dungeons
-    elif val in range(21612, 21618):
-        return "Technical"
-    elif val in range(38901, 38912):
-        return "Descended (original)"
-    elif val in range(200101, 200111):
-        return "Alt. Technial"
-    elif val in range(200021, 200057):
-        return "Technical"
-    elif val in range(200301, 200306) or val in range(200201, 200206):
-        return "Special Decended"
-    elif val in DUNGEON_TYPE_DESCRIPTORS:
-        return DUNGEON_TYPE_DESCRIPTORS[val]
-    else:
-        return "No Data"
