+++ pad_api_data/pad_etl/common/pad_util.py
@@ -1,6 +1,7 @@
 import datetime
 import json
 import re
+from .dungeon_types import DUNGEON_TYPE_COMMENTS
 
 
 def strip_colors(message: int) -> str:
@@ -60,3 +61,23 @@ class JsonDictEncodable(json.JSONEncoder):
 
     def default(self, o):
         return o.__dict__
+
+# Simply returns a dungeon comment for a input raw value. Doing it this way is necessary as to not have to put all the values
+# directly into a dictionary when multiple val's correspond to a single comment, but are unnecessarily delineated
+def get_dungeon_comment(val):
+    if val in range(5611, 5615):
+        return "Retired Special Dungeons"  # These are the last normal dungeons
+    elif val in range(21612, 21618):
+        return "Technical"
+    elif val in range(38901, 38912):
+        return "Descended (original)"
+    elif val in range(200101, 200111):
+        return "Alt. Technial"
+    elif val in range(200021, 200057):
+        return "Technical"
+    elif val in range(200301, 200306) or val in range(200201, 200206):
+        return "Special Decended"
+    elif val in DUNGEON_TYPE_COMMENTS:
+        return DUNGEON_TYPE_COMMENTS[val]
+    else:
+        return "No Data"
