+++ pad_api_data/pad_etl/data/dungeon.py
@@ -9,7 +9,7 @@ import os
 from typing import List, Any
 
 from ..common import pad_util
-from ..common.dungeon_types import DUNGEON_TYPE, get_dungeon_comment
+from ..common.dungeon_types import DUNGEON_TYPE
 
 
 # The typical JSON file name for this data.
@@ -35,18 +35,6 @@ class DungeonFloor(pad_util.JsonDictEncodable):
         self.unknown_010 = raw[10]
 
 
-prefix_to_dungeontype = {
-    # #G#Ruins of the Star Vault 25
-    '#G#': 'guerrilla',
-
-    # #1#Star Treasure of the Night Sky 25
-    '#1#': 'unknown-1',
-
-    # #C#Rurouni Kenshin dung
-    '#C#': 'collab',
-}
-
-
 class Dungeon(pad_util.JsonDictEncodable):
     """A top-level dungeon."""
 
@@ -66,17 +54,7 @@ class Dungeon(pad_util.JsonDictEncodable):
 
         # I call it comment as it is similar to dungeon_type, but sometimes designates certain dungeons specifically
         # over others. See dungeon_types.py for more details.
-        self.dungeon_comment = get_dungeon_comment(int(raw[5]))
-
-
-        self.prefix = None  # type: str
-
-        for prefix, dungeon_type in prefix_to_dungeontype.items():
-            if self.clean_name.startswith(prefix):
-                self.prefix = prefix
-                self.dungeon_type = dungeon_type
-                self.clean_name = self.clean_name[len(prefix):]
-                break
+        self.dungeon_comment = pad_util.get_dungeon_comment(int(raw[5]))
 
         if len(raw) > 6:
             print('unexpected field count: ' + ','.join(raw))
